feat: extract gallery grid into its own component to reuse
