+++ apps/mobile/src/app/(tabs)/profile/_layout.tsx
@@ -15,6 +15,15 @@ export default function ProfileLayout() {
       <Stack.Screen name="friend-requests" options={{ title: 'Friend Requests' }} />
       <Stack.Screen name="settings" options={{ title: 'Settings' }} />
       <Stack.Screen name="edit" options={{ title: 'Edit Profile' }} />
+      <Stack.Screen
+        name="gallery-picker"
+        options={{
+          title: 'Gallery',
+          presentation: 'modal',
+          animation: 'slide_from_bottom',
+          animationDuration: 280,
+        }}
+      />
       <Stack.Screen name="media-feed" options={{ title: 'Media' }} />
     </Stack>
   );
